Moved associate button to the file explorer pane. Associate button can now associate to the root node of the classification scheme.
